- Cycle cleanup; - bug in now (wrong month) fixed; - off-by-one in Psalms Chitas fixed.
