+++ library/src/test/scala/org/podval/calendar/astronomy/TextTest.scala
@@ -5,7 +5,7 @@ import org.podval.calendar.angles.Angles
 import Angles.{Position, Rotation, headRange, range}
 import org.podval.calendar.jewish.{Jewish, LeapYearsCycle}
 import Jewish.{Day, Month, Year}
-import org.podval.calendar.dates.Cycle
+import org.podval.calendar.dates.YearsCycle
 import org.podval.calendar.numbers.BigRational
 
 class TextTest extends FlatSpec with Matchers {
@@ -39,7 +39,7 @@ class TextTest extends FlatSpec with Matchers {
 
   "epoch" should "be as in KH 11:16" in {
     LeapYearsCycle.inCycle(260, 17) shouldBe 4938
-    LeapYearsCycle.forNumber(4938) shouldBe Cycle.In(260, 17)
+    LeapYearsCycle.forNumber(4938) shouldBe YearsCycle.In(260, 17)
     Epoch.Text.day shouldBe Year(4938).month(Month.Name.Nisan).day(3)
     Epoch.Text.day.name shouldBe Day.Name.Chamishi
   }
